+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## [0.0.7] - TODO: more features.
+## [0.0.8] - TODO: more features.
 
 * 0.0.1+1: Add example.
 * 0.0.1+2: Add more Examples and add rebuild method in Model.
@@ -6,3 +6,10 @@
 * 0.0.1+4: Change search parameter to params & add new parameter #path for get part of json & Add controller in model generator
 * 0.0.1+5: fix bugs & add new item McController for add/get/remove your model Now you can use your model from another screen without traditional way
 * 0.0.1+6: Optimization (add) method of McController & possible to add McRequest in controller
+* 0.0.1+7: 
+- fix bugs
+- Inject [mc] (instance) of McController in Stateless and ful widget by extension for use it easily 
+- remove path parameter
+- add complex parameter for complex json
+- add inspect parameter for complex json if complex true you need to define inspect by function return List of Map like this [{'item':'value1'},{'item':'value1'}] & if data already like this you dont need define complex and inspect
+- add sendFile Method for send files